requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
pyyaml>=6.0.1
requests>=2.31.0
pybaseball>=2.2.7
zoneinfo; python_version < "3.9"
